Decrement positive sensor count by 1
It seems as though the bridge responds with a sensor count that is one higher than the actual number of sensors when any sensors are paired. If no sensors are paired, it responds with 0. Decrement the positive sensor count by 1 so we don't wait around until the timeout value is reached for the n+1'th sensor MAC to come back.