Update README.md
I had to re-install the wyzesense integration due to an error with HomeKit. Without removing the hub from my Home Assistant device first, the setup would fail over and over. Removing it and following these instructions exactly made the integration install correctly. This note may be helpful for others.
Also changed spelling to match the rest of the document.